`react` renamed from `react-internal` (consistency in `eslint-config`)
