+++ packages/eslint-config/next.ts
@@ -1,6 +1,6 @@
 import tseslint, { type ConfigArray } from "typescript-eslint";
 import pluginNext from "@next/eslint-plugin-next";
-import { reactConfig } from "./react-internal.js";
+import { reactConfig } from "./react.js";
 
 export const nextJsConfig: ConfigArray = tseslint.config([
   // @ts-ignore
