+++ packages/eslint-config/package.json
@@ -6,7 +6,7 @@
   "exports": {
     "./base": "./base.ts",
     "./next-js": "./next.ts",
-    "./react-internal": "./react-internal.ts"
+    "./react": "./react.ts"
   },
   "devDependencies": {
     "@eslint/js": "^9.31.0",
