+++ packages/ui/eslint.config.ts
@@ -1,4 +1,4 @@
-import { reactConfig } from "@repo/eslint-config/react-internal";
+import { reactConfig } from "@repo/eslint-config/react";
 import { type ConfigArray } from "typescript-eslint";
 
 const config: ConfigArray = reactConfig(true);
